+++ src/main/java/com/mycompany/webapp/controller/ReservationController.java
@@ -10,9 +10,24 @@ import lombok.extern.log4j.Log4j2;
 @RequestMapping("/reservation")
 public class ReservationController {
 	
-	@RequestMapping("/")//상단 탭 진료 예약 클릭시, 예약화면으로 이동(영화관 예약 비슷한 그 화면...)
-	public String reservation() {
+	//header의 탭 진료 예약 클릭시, 예약화면으로 이동
+	@RequestMapping("/reservationMain")
+	public String reservationMain() {
 		log.info("실행");
-		return "reservation/reservation";
+		return "reservation/reservationMain";
+	}
+	
+	//reservationMain화면에서 특정 치과 1개 클릭한 경우의 치과 상세 페이지.
+	@RequestMapping("/dentistDetail")
+	public String dentistDetail() {
+		log.info("실행");
+		return "reservation/dentistDetail";
+	}
+	
+	//특정 치과 상세 페이지에서 '예약하기' 버튼 클릭시에 달력과 버튼을 이용한 예약신청 화면으로 이동.
+	@RequestMapping("/reservationUsingCalendar")
+	public String reservationUsingCalendar() {
+		log.info("실행");
+		return "reservation/reservationUsingCalendar";
 	}
 }
